feature: criar classe model produto
